fix!: changed types to match a more direct input

+++ structs/external-api-responses/wheater-forecast.go
@@ -4,6 +4,6 @@ type WheaterForecast struct {
     Hourly struct {
         Time []string `json:"time"`
         Temperature []float64 `json:"temperature_2m"`
-        Humidity []int8`json:"relativehumidity_2m"`
+        Humidity []float64 `json:"relativehumidity_2m"`
     } `json:"hourly"`
 }
